# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentDialog Agent API",
    "version": "dev",
    "description": "The surface an AI agent integrates against: conversations, messages, human queries and webhooks."
  },
  "servers": [
    {
      "url": "https://api.agentdialog.io"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/agent/conversations": {
      "post": {
        "tags": [
          "conversations"
        ],
        "summary": "Start a conversation",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Repeat a POST with the same key and the original response comes back instead of the work happening twice. Only successful responses are remembered, so a refusal leaves the key free for the corrected retry.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "maxLength": 256
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2048
                  },
                  "context": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "intentType": {
                    "type": "string",
                    "enum": [
                      "permission",
                      "clarification",
                      "solicitation",
                      "notification"
                    ]
                  },
                  "settings": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The conversation, created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "createdByAgentId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "title": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "description": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "active",
                            "archived",
                            "closed"
                          ]
                        },
                        "context": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "intentType": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "enum": [
                            "permission",
                            "clarification",
                            "solicitation",
                            "notification",
                            null
                          ]
                        },
                        "settings": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "createdByAgentId",
                        "title",
                        "description",
                        "status",
                        "context",
                        "intentType",
                        "settings",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/IdempotencyKeyReused"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "422": {
            "description": "The request body failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "get": {
        "tags": [
          "conversations"
        ],
        "summary": "List your conversations",
        "description": "Paginated, but not with the API's usual envelope: this route builds its own pagination object with only hasMore and count, and never returns nextCursor or prevCursor even though the cursor query param is accepted.",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversations the authenticated agent created, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "createdByAgentId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "title": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "description": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "active",
                              "archived",
                              "closed"
                            ]
                          },
                          "context": {
                            "type": "object",
                            "additionalProperties": {}
                          },
                          "intentType": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "enum": [
                              "permission",
                              "clarification",
                              "solicitation",
                              "notification",
                              null
                            ]
                          },
                          "settings": {
                            "type": "object",
                            "additionalProperties": {}
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "createdByAgentId",
                          "title",
                          "description",
                          "status",
                          "context",
                          "intentType",
                          "settings",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "hasMore": {
                          "type": "boolean"
                        },
                        "count": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "hasMore",
                        "count"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "description": "The `limit` or `cursor` query parameter failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/v1/agent/conversations/{id}/invitations": {
      "post": {
        "tags": [
          "invitations"
        ],
        "summary": "Invite a human to a conversation",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Repeat a POST with the same key and the original response comes back instead of the work happening twice. Only successful responses are remembered, so a refusal leaves the key free for the corrected retry.",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "message": {
                    "type": "string",
                    "maxLength": 1024
                  },
                  "language": {
                    "type": "string",
                    "enum": [
                      "en",
                      "es",
                      "ca"
                    ]
                  },
                  "expiresInHours": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 168,
                    "default": 48
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The invitation, created — or, if this agent already has a prior accepted invitation with this human and no revoked trust since, auto-accepted immediately (`autoAccepted: true`, `status: \"accepted\"`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "conversationId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "invitedByAgentId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "invitedHumanEmail": {
                          "type": "string"
                        },
                        "token": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "pending",
                            "accepted",
                            "declined",
                            "expired",
                            "revoked"
                          ]
                        },
                        "message": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "language": {
                          "type": "string",
                          "enum": [
                            "en",
                            "es",
                            "ca"
                          ]
                        },
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "autoAccepted": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "conversationId",
                        "invitedByAgentId",
                        "invitedHumanEmail",
                        "token",
                        "status",
                        "message",
                        "language",
                        "expiresAt",
                        "createdAt",
                        "updatedAt",
                        "autoAccepted"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "The authenticated agent is not a participant in this conversation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "A pending invitation to this email already exists for this conversation. May also indicate a reused `Idempotency-Key`; the `code` field distinguishes them.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "422": {
            "description": "The request body failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "get": {
        "tags": [
          "invitations"
        ],
        "summary": "List invitations for a conversation",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Invitations sent for this conversation, in any status.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "conversationId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "invitedByAgentId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "invitedHumanEmail": {
                            "type": "string"
                          },
                          "token": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "pending",
                              "accepted",
                              "declined",
                              "expired",
                              "revoked"
                            ]
                          },
                          "message": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "language": {
                            "type": "string",
                            "enum": [
                              "en",
                              "es",
                              "ca"
                            ]
                          },
                          "expiresAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "conversationId",
                          "invitedByAgentId",
                          "invitedHumanEmail",
                          "token",
                          "status",
                          "message",
                          "language",
                          "expiresAt",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "The authenticated agent is not a participant in this conversation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/agent/conversations/invitations/{id}": {
      "delete": {
        "tags": [
          "invitations"
        ],
        "summary": "Revoke a pending invitation",
        "description": "id here is the invitation id, not a conversation id.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The invitation, revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "conversationId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "invitedByAgentId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "invitedHumanEmail": {
                          "type": "string"
                        },
                        "token": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "pending",
                            "accepted",
                            "declined",
                            "expired",
                            "revoked"
                          ]
                        },
                        "message": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "language": {
                          "type": "string",
                          "enum": [
                            "en",
                            "es",
                            "ca"
                          ]
                        },
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "conversationId",
                        "invitedByAgentId",
                        "invitedHumanEmail",
                        "token",
                        "status",
                        "message",
                        "language",
                        "expiresAt",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Only a pending invitation can be revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "No such invitation, or it wasn't created by the authenticated agent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "An agent API key, prefixed `mge_ag_`."
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "NOT_FOUND",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "CONFLICT",
                  "VALIDATION_ERROR",
                  "RATE_LIMIT",
                  "IDEMPOTENCY_IN_PROGRESS",
                  "IDEMPOTENCY_KEY_REUSED",
                  "UNDECIDABLE_QUERY",
                  "PAYLOAD_TOO_LARGE",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": {}
              },
              "retryAfter": {
                "type": "integer"
              },
              "reason": {
                "type": "string"
              },
              "detail": {
                "type": "string"
              },
              "remedy": {
                "type": "string"
              },
              "prior_query_id": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "The bearer token is missing, malformed, or does not match a live agent.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Too many requests. `retryAfter` in the error envelope is the number of seconds to wait before trying again.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "IdempotencyKeyReused": {
        "description": "The Idempotency-Key header was reused, either by a request still in flight or by one whose original request had a different body.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "The request body exceeded the server's configured maximum size (see app.ts's bodyLimit).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "InternalError": {
        "description": "An unexpected error. The error envelope carries no further detail than `message`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    }
  }
}